Actors and their shared events

Given an affiliation network and a vector of actor node IDs, produce the induced subgraph on the actor nodes together with all event nodes incident to at least two of them. This is called the actors' schedule.

Usage

schedule(graph, actors = V(graph)[V(graph)$type == FALSE])

Arguments

graph

An affiliation network.

actors

A vector of actor nodes in graph.

See also